Rename tagDetails ad tagLabels
