--- src/routes/settings/tagDetails/index.ts
@@ -1,1 +1,0 @@
-export { default as TagDetails } from './tagDetails';
