+++ src/routes/settings/tagLabels/index.ts
@@ -1,0 +1,1 @@
+export { default as TagLabels } from './tagLabels';
